Spareted styles in a separete file
