+++ react-native-app/styles.js
@@ -1,0 +1,14 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  container: {
+    marginTop: 25,
+    padding: 10,
+  },
+  nav: {
+    flexDirection: "row",
+    justifyContent: "space-around",
+  },
+});
+
+export default styles;
